Refactor scss
