--- src/lib/mapveu/analysis/SemiTransparentBanner.scss
@@ -1,87 +1,0 @@
-.SemiTransparentBanner {
-  align-items: center;
-  background: rgba(255, 255, 255, 0.8);
-  display: flex;
-  min-height: 110px;
-  justify-content: space-between;
-  position: absolute;
-  right: 0;
-  top: 0;
-  transition: all 0.1s ease;
-  width: 100%;
-  z-index: 10;
-
-  &__Contents {
-    align-items: center;
-    display: flex;
-    justify-content: flex-start;
-    margin-bottom: 0.5rem;
-    width: 75%;
-  }
-
-  &__collapsed {
-    background: none;
-    top: -100px;
-  }
-
-  .ExpandToggleContainer {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    position: absolute;
-    bottom: -20px;
-    left: 50%;
-    transform: translate(-50%, 0%);
-
-    &--collapsed {
-      position: absolute;
-      bottom: 20px;
-      background: none;
-    }
-
-    &__HoverIndicator {
-      width: 100%;
-      height: 3px;
-      background: transparent;
-      position: absolute;
-      bottom: 0;
-    }
-  }
-
-  .ExpandToggleButton {
-    z-index: 100;
-    margin-right: 0;
-    display: flex;
-    justify-content: center;
-    align-items: flex-end;
-    background: rgba(255, 255, 255, 0.8);
-    border: none;
-    width: 60px;
-    height: 40px;
-    position: absolute;
-
-    &__collapsed {
-      width: 60px;
-      margin-top: 50px;
-    }
-  }
-
-  .SvgContainer {
-    border: none;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    width: 60px;
-    height: 40px;
-    fill: #000;
-  }
-  .ArrowIcon {
-    transform: rotate(-90deg);
-    width: 30px;
-    height: 40px;
-
-    &__collapsed {
-      transform: rotate(90deg);
-    }
-  }
-}
